requirements.txt
faiss-cpu==1.9.0
numpy==1.26.4
sentence-transformers==3.2.1